+++ packages/moonstone/Input/tests/Input-specs.js
@@ -2,9 +2,10 @@ import React from 'react';
 import {mount} from 'enzyme';
 import sinon from 'sinon';
 import Input from '../Input';
+import Spotlight from '@enact/spotlight';
 
 describe('Input Specs', () => {
-	it('Should have an input element', function () {
+	it('should have an input element', function () {
 		const subject = mount(
 			<Input />
 		);
@@ -12,7 +13,7 @@ describe('Input Specs', () => {
 		expect(subject.find('input')).to.have.length(1);
 	});
 
-	it('Should include a placeholder if specified', function () {
+	it('should include a placeholder if specified', function () {
 		const subject = mount(
 			<Input placeholder="hello" />
 		);
@@ -20,7 +21,7 @@ describe('Input Specs', () => {
 		expect(subject.find('input').prop('placeholder')).to.equal('hello');
 	});
 
-	it('Should callback onChange when the text changes', function () {
+	it('should callback onChange when the text changes', function () {
 		const handleChange = sinon.spy();
 		const value = 'blah';
 		const evt = {target: {value: value}};
@@ -36,7 +37,7 @@ describe('Input Specs', () => {
 		expect(actual).to.equal(expected);
 	});
 
-	it('Should blur input on enter if dismissOnEnter', function () {
+	it('should blur input on enter if dismissOnEnter', function () {
 		const node = document.body.appendChild(document.createElement('div'));
 		const handleChange = sinon.spy();
 
@@ -56,7 +57,7 @@ describe('Input Specs', () => {
 		expect(actual).to.equal(expected);
 	});
 
-	it('Should be able to be disabled', function () {
+	it('should be able to be disabled', function () {
 		const subject = mount(
 			<Input disabled />
 		);
@@ -64,7 +65,7 @@ describe('Input Specs', () => {
 		expect(subject.find('input').prop('disabled')).to.true();
 	});
 
-	it('Should reflect the value if specified', function () {
+	it('should reflect the value if specified', function () {
 		const subject = mount(
 			<Input value="hello" />
 		);
@@ -72,7 +73,7 @@ describe('Input Specs', () => {
 		expect(subject.find('input').prop('value')).to.equal('hello');
 	});
 
-	it('Should have dir equal to rtl when there is rtl text', function () {
+	it('should have dir equal to rtl when there is rtl text', function () {
 		const subject = mount(
 			<Input value="שועל החום הזריז קפץ מעל הכלב העצלן.ציפור עפה השעועית עם שקי" />
 		);
@@ -83,7 +84,7 @@ describe('Input Specs', () => {
 		expect(actual).to.equal(expected);
 	});
 
-	it('Should have dir equal to ltr when there is ltr text', function () {
+	it('should have dir equal to ltr when there is ltr text', function () {
 		const subject = mount(
 			<Input value="content" />
 		);
@@ -94,7 +95,7 @@ describe('Input Specs', () => {
 		expect(actual).to.equal(expected);
 	});
 
-	it('Should have dir equal to rtl when there is rtl text in the placeholder', function () {
+	it('should have dir equal to rtl when there is rtl text in the placeholder', function () {
 		const subject = mount(
 			<Input value="שועל החום הזריז קפץ מעל הכלב העצלן.ציפור עפה השעועית עם שקי" />
 		);
@@ -105,7 +106,7 @@ describe('Input Specs', () => {
 		expect(actual).to.equal(expected);
 	});
 
-	it('Should have dir equal to ltr when there is ltr text in the placeholder', function () {
+	it('should have dir equal to ltr when there is ltr text in the placeholder', function () {
 		const subject = mount(
 			<Input placeholder="content" />
 		);
@@ -116,7 +117,7 @@ describe('Input Specs', () => {
 		expect(actual).to.equal(expected);
 	});
 
-	it('Should have dir equal to rtl when there is ltr text in the placeholder, but rtl text in value', function () {
+	it('should have dir equal to rtl when there is ltr text in the placeholder, but rtl text in value', function () {
 		const subject = mount(
 			<Input
 				placeholder="content"
@@ -130,7 +131,7 @@ describe('Input Specs', () => {
 		expect(actual).to.equal(expected);
 	});
 
-	it('Should have dir equal to ltr when there is rtl text in the placeholder, but ltr text in value', function () {
+	it('should have dir equal to ltr when there is rtl text in the placeholder, but ltr text in value', function () {
 		const subject = mount(
 			<Input
 				placeholder="שועל החום הזריז קפץ מעל הכלב העצלן.ציפור עפה השעועית עם שקי"
@@ -141,6 +142,37 @@ describe('Input Specs', () => {
 		const expected = 'ltr';
 		const actual = subject.find('input').prop('dir');
 
+		expect(actual).to.equal(expected);
+	});
+
+	it('Should pause spotlight when input has focus', function () {
+		const pauseSpy = sinon.spy(Spotlight, 'pause');
+		const subject = mount(
+			<Input />
+		);
+
+		subject.simulate('click');
+
+		const expected = true;
+		const actual = pauseSpy.calledOnce;
+
+		Spotlight.pause.restore();
+		expect(actual).to.equal(expected);
+	});
+
+	it('Should resume spotlight on unmount', function () {
+		const resumeSpy = sinon.spy(Spotlight, 'resume');
+		const subject = mount(
+			<Input />
+		);
+
+		subject.simulate('click');
+		subject.unmount();
+
+		const expected = true;
+		const actual = resumeSpy.calledOnce;
+
+		Spotlight.resume.restore();
 		expect(actual).to.equal(expected);
 	});
 });