+++ packages/moonstone/VirtualList/tests/VirtualList-specs.js
@@ -2,7 +2,6 @@ import {mount} from 'enzyme';
 import React from 'react';
 
 import Item from '../../Item';
-
 import VirtualList from '../VirtualList';
 
 describe('VirtualList Specs', () => {
@@ -44,28 +43,28 @@ describe('VirtualList Specs', () => {
 		);
 
 		describe('Set props Specs', () => {
-			it('Should render a list item of \'data\'', function () {
+			it('should render a list item of \'data\'', function () {
 				const expected = 'Account 0';
 				const actual = subject.find('VirtualListCore').children().at(0).text();
 
 				expect(actual).to.equal(expected);
 			});
 
-			it('Should have the prop \'dataSize\' of 100', function () {
+			it('should have the prop \'dataSize\' of 100', function () {
 				const expected = 100;
 				const actual = subject.find('VirtualListCore').prop('dataSize');
 
 				expect(actual).to.equal(expected);
 			});
 
-			it('Should have the prop \'direction\' of \'horizontal\'', function () {
+			it('should have the prop \'direction\' of \'horizontal\'', function () {
 				const expected = 'horizontal';
 				const actual = subject.find('VirtualListCore').prop('direction');
 
 				expect(actual).to.equal(expected);
 			});
 
-			it('Should apply list size', function () {
+			it('should apply list size', function () {
 				const expectedWidth = '500px';
 				const expectedHeight = '700px';
 				const actualWidth = subject.prop('style').width;
@@ -75,14 +74,14 @@ describe('VirtualList Specs', () => {
 				expect(actualHeight).to.equal(expectedHeight);
 			});
 
-			it('Should apply background color', function () {
+			it('should apply background color', function () {
 				const expected = 'red';
 				const actual = subject.prop('style').backgroundColor;
 
 				expect(actual).to.equal(expected);
 			});
 
-			it('Should scroll with cbScrollTo prop', function () {
+			it('should scroll with cbScrollTo prop', function () {
 				myScrollTo({position: {x: 10}, animate: false});
 
 				const expected = 10;
@@ -93,7 +92,7 @@ describe('VirtualList Specs', () => {
 		});
 
 		describe('Change props Specs', () => {
-			it('Should change value of the prop \'data\' to \'Password 0\'', function () {
+			it('should change value of the prop \'data\' to \'Password 0\'', function () {
 				data[0] = {name: 'Password 0'};
 				subject.setProps({data: data});
 
@@ -103,7 +102,7 @@ describe('VirtualList Specs', () => {
 				expect(actual).to.equal(expected);
 			});
 
-			it('Should change value of the prop \'direction\' to \'horizontal\'', function () {
+			it('should change value of the prop \'direction\' to \'horizontal\'', function () {
 				subject.setProps({direction: 'vertical'});
 
 				const expected = 'vertical';
@@ -112,7 +111,7 @@ describe('VirtualList Specs', () => {
 				expect(actual).to.equal(expected);
 			});
 
-			it('Should change value of the prop \'spacing\' to 5', function () {
+			it('should change value of the prop \'spacing\' to 5', function () {
 				subject.setProps({spacing: 5});
 
 				const expected = 5;
@@ -121,7 +120,7 @@ describe('VirtualList Specs', () => {
 				expect(actual).to.equal(expected);
 			});
 
-			it('Should change value of the props \'style\' to \'backgroundColor\', \'width\', and \'height\'', function () {
+			it('should change value of the props \'style\' to \'backgroundColor\', \'width\', and \'height\'', function () {
 				subject.setProps({style: {backgroundColor: 'blue', width: '400px', height: '600px'}});
 
 				const expectedBackgroundColor = 'blue';