+++ packages/spotlight/CHANGELOG.md
@@ -7,6 +7,9 @@ The following is a curated list of changes in the Enact spotlight module, newest
 ### Added
 
 - `@enact/spotlight.spottableClass` export
+- `spotlightDisabled` prop support in `@enact/spotlight/container`
+- `onSpotlightDown`, `onSpotlightLeft`, `onSpotlightRight`, and `onSpotlightUp` prop support in `@enact/spotlight/spottable`
+- `onSpotlightDisappear` event property support in `@enact/spotlight/spottable`
 
 ### Removed